fix soda player issues

+++ src/electron/sodaPlayer.ts
@@ -4,10 +4,14 @@ import execa from "execa";
 import fs from "fs";
 import got, { Progress } from "got";
 import path from "path";
+import rifraf from "rimraf";
+import { patchSodaPlayer, sodaPlayerBasicConfig } from "soda-player-patch";
+import { getPaths } from "soda-player-patch/build/patchElectronApp";
 import stream from "stream";
 import { typedIpcMain } from "typed-ipc";
 import util from "util";
 
+import { settingsStore } from "../react/electron-shared/settings";
 import { mainWindow } from "./mainWindow";
 
 const pipeline = util.promisify(stream.pipeline);
@@ -23,43 +27,67 @@ export const sodaPlayerExecPath = path.join(process.env.LOCALAPPDATA || "", "sod
 export const isSodaPlayerInstalled = (): boolean =>
     fs.existsSync(sodaPlayerExecPath);
 
-export const installSodaPlayer = async () => {
-    const sodaPlayerSavePath = path.join(app.getPath("temp"), "filmomania-soda-player.exe");
+export const installOrAndPatchSodaPlayer = async () => {
+    if (!isSodaPlayerInstalled()) {
+        const sodaPlayerSavePath = path.join(app.getPath("temp"), "filmomania-soda-player.exe");
 
-    if (process.platform !== "win32") {
-        //todo-moderate
-        throw new Error("Only win");
+        if (process.platform !== "win32") {
+            //todo-moderate
+            throw new Error("Only win");
+        }
+
+        await pipeline(
+            got.stream(sodaPlayerDownloadUrl)
+                .on("downloadProgress", (progress: Progress) => {
+                    typedIpcMain.sendToWindow(mainWindow, "updateSodaPlayerInstallationState", {
+                        stage: "downloading",
+                        progress: progress.percent,
+                        downloadedBytes: progress.transferred
+                    });
+                }),
+            fs.createWriteStream(sodaPlayerSavePath)
+        );
+
+        typedIpcMain.sendToWindow(mainWindow, "updateSodaPlayerInstallationState", {
+            stage: "installing"
+        });
+        await execa(sodaPlayerSavePath, ["/s"]);
+        // todo-moderate
+        fs.unlinkSync(sodaPlayerSavePath);
     }
-
-    await pipeline(
-        got.stream(sodaPlayerDownloadUrl)
-            .on("downloadProgress", (progress: Progress) => {
-                typedIpcMain.sendToWindow(mainWindow, "updateSodaPlayerInstallationState", {
-                    stage: "downloading",
-                    progress: progress.percent,
-                    downloadedBytes: progress.transferred
-                });
-            }),
-        fs.createWriteStream(sodaPlayerSavePath)
-    );
-
     typedIpcMain.sendToWindow(mainWindow, "updateSodaPlayerInstallationState", {
-        stage: "installing"
+        stage: "patching"
     });
-    await execa(sodaPlayerSavePath, ["/s"]);
-    // todo-moderate
-    fs.unlinkSync(sodaPlayerSavePath);
-    // todo modernaze
+    let patched = false;
+    try {
+        await patchSodaPlayer();
+        patched = true;
+    } catch (err) {
+        console.error(err);
+        // something went wrong while patching. try to restore original version
+        const { oldAsarSource, asarUnpacked, asarSource } = await getPaths(sodaPlayerBasicConfig.localAppdataDir);
+        // throw fix ipc
+        // patch was complete (but probably something broke)
+        if (fs.existsSync(oldAsarSource)) {
+            if (fs.existsSync(asarUnpacked)) {
+                rifraf.sync(asarUnpacked);
+            }
+            await fs.promises.rename(oldAsarSource, asarSource);
+        }
+    }
     typedIpcMain.sendToWindow(mainWindow, "updateSodaPlayerInstallationState", {
-        stage: "installed"
+        stage: "done",
+        patched
     });
 };
 
 export const playWithSodaPlayer = async (magnet: string) => {
+    console.log("run", isSodaPlayerInstalled(), settingsStore.get("generalDefaultPlayer"));
     if (!isSodaPlayerInstalled()) return;
+    const defaultPlayer = settingsStore.get("generalDefaultPlayer") as string;
     // todo-high check arg
     // todo-very-high
-    child_process.spawn(`C:\\Users\\Professional\\AppData\\Local\\sodaplayer\\Soda Player.exe`, [magnet], {
+    child_process.spawn(`C:\\Users\\Professional\\AppData\\Local\\sodaplayer\\Soda Player.exe`, [magnet, ...defaultPlayer === "sodaPlayerPatched" ? ["--fullscreen"] : []], {
         detached: true,
         stdio: "ignore"
     });
